+++ tests/test_spider.py
@@ -1,6 +1,8 @@
 # coding=utf-8
 
-from xpaw.spider import SpiderMiddlewareManager
+import pytest
+
+from xpaw.spider import SpiderMiddlewareManager, Spider
 from xpaw.eventbus import EventBus
 from xpaw.config import Config
 from xpaw import events
@@ -85,18 +87,22 @@ async def test_spider_middleware_manager_handlers():
     spidermw = SpiderMiddlewareManager.from_cluster(cluster)
     response_obj = object()
     result_obj = object()
+    error_obj = object()
     await cluster.event_bus.send(events.cluster_start)
     await spidermw._handle_start_requests((result_obj,))
     await spidermw._handle_input(response_obj)
     await spidermw._handle_output(response_obj, (result_obj,))
+    await spidermw._handle_error(response_obj, error_obj)
     await cluster.event_bus.send(events.cluster_shutdown)
     assert 'open' in data and 'close' in data
     assert data['handle_start_requests'] is result_obj
     assert data['handle_input'] is response_obj
     assert data['handle_output'][0] is response_obj and data['handle_output'][1] is result_obj
+    assert data['handle_error'][0] is response_obj and data['handle_error'][1] is error_obj
     assert data['async_handle_start_requests'] is result_obj
     assert data['async_handle_input'] is response_obj
     assert data['async_handle_output'][0] is response_obj and data['async_handle_output'][1] is result_obj
+    assert data['async_handle_error'][0] is response_obj and data['async_handle_error'][1] is error_obj
 
     data2 = {}
     cluster2 = Cluster(spider_middlewares={lambda d=data2: MySpidermw(d): 0},
@@ -112,3 +118,30 @@ async def test_spider_middleware_manager_handlers():
 
     cluster3 = Cluster(spider_middlewares=None, spider_middlewares_base=None, data={})
     SpiderMiddlewareManager.from_cluster(cluster3)
+
+
+class FooSpider(Spider):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.data = self.config['data']
+
+    def open(self):
+        super().open()
+        self.data['open'] = ''
+
+    def close(self):
+        super().close()
+        self.data['close'] = ''
+
+
+async def test_spider(loop):
+    data = {}
+    cluster = Cluster(data=data, loop=loop)
+    spider = FooSpider.from_cluster(cluster)
+    await cluster.event_bus.send(events.cluster_start)
+    with pytest.raises(NotImplementedError):
+        spider.start_requests()
+    with pytest.raises(NotImplementedError):
+        spider.parse(None)
+    await cluster.event_bus.send(events.cluster_shutdown)
+    assert 'open' in data and 'close' in data