+++ xpaw/spider.py
@@ -2,10 +2,11 @@
 
 import logging
 import inspect
+from asyncio import CancelledError
 
 from .middleware import MiddlewareManager
 from . import events
-from .utils import AsyncGenWrapper
+from .utils import AsyncGenWrapper, iterable_to_list
 
 log = logging.getLogger(__name__)
 
@@ -50,6 +51,7 @@ class SpiderMiddlewareManager(MiddlewareManager):
     def __init__(self, *middlewares):
         self._input_handlers = []
         self._output_handlers = []
+        self._error_handlers = []
         self._start_requests_handlers = []
         super().__init__(*middlewares)
 
@@ -59,6 +61,8 @@ class SpiderMiddlewareManager(MiddlewareManager):
             self._input_handlers.append(middleware.handle_input)
         if hasattr(middleware, "handle_output"):
             self._output_handlers.insert(0, middleware.handle_output)
+        if hasattr(middleware, "handle_error"):
+            self._error_handlers.insert(0, middleware.handle_error)
         if hasattr(middleware, "handle_start_requests"):
             self._start_requests_handlers.insert(0, middleware.handle_start_requests)
 
@@ -67,28 +71,47 @@ class SpiderMiddlewareManager(MiddlewareManager):
         return cls._make_component_list('spider_middlewares', config)
 
     async def parse(self, spider, response):
-        await self._handle_input(response)
-        if response.request and response.request.callback:
-            r = getattr(spider, response.request.callback)(response)
-        else:
-            r = spider.parse(response)
-        if inspect.iscoroutine(r):
-            r = await r
-        assert r is None or _isiterable(r), \
-            "Result of parsing must be None or an iterable object, got {}".format(type(r).__name__)
-        if r:
-            r = await self._handle_output(response, r)
-        if r is None:
-            r = ()
-        if not hasattr(r, "__aiter__"):
-            r = AsyncGenWrapper(r)
-        return r
+        request = response.request
+        try:
+            try:
+                await self._handle_input(response)
+            except CancelledError:
+                raise
+            except Exception as e:
+                await self.handle_error(spider, request, e)
+                raise e
+            if request.callback:
+                res = getattr(spider, request.callback)(response)
+            else:
+                res = spider.parse(response)
+            if inspect.iscoroutine(res):
+                res = await res
+            assert res is None or _isiterable(res), \
+                "Parsing result must be None or an iterable object, got {}".format(type(result).__name__)
+            result = await iterable_to_list(res)
+        except CancelledError:
+            raise
+        except Exception as e:
+            res = await self._handle_error(response, e)
+            if isinstance(res, Exception):
+                raise res
+            result = await iterable_to_list(res)
+        if result:
+            res = await self._handle_output(response, result)
+            result = await iterable_to_list(res)
+        return result
 
-    async def handle_error(self, spider, request, error):
-        if request and request.errback:
-            r = getattr(spider, request.errback)(request, error)
-            if inspect.iscoroutine(r):
-                await r
+    @staticmethod
+    async def handle_error(spider, request, error):
+        try:
+            if request and request.errback:
+                r = getattr(spider, request.errback)(request, error)
+                if inspect.iscoroutine(r):
+                    await r
+        except CancelledError:
+            raise
+        except Exception:
+            log.warning("Error occurred in error callback", exc_info=True)
 
     async def start_requests(self, spider):
         r = spider.start_requests()
@@ -118,6 +141,17 @@ class SpiderMiddlewareManager(MiddlewareManager):
             assert _isiterable(result), \
                 "Output handler must return an iterable object, got {}".format(type(result).__name__)
         return result
+
+    async def _handle_error(self, response, error):
+        for method in self._error_handlers:
+            res = method(response, error)
+            if inspect.iscoroutine(res):
+                res = await res
+            assert res is None or _isiterable(res), \
+                "Exception handler must return None or an iterable object, got {}".format(type(res).__name__)
+            if res is not None:
+                return res
+        return error
 
     async def _handle_start_requests(self, result):
         for method in self._start_requests_handlers: